Move get_available_nodes callback to package manager
